refactor(aruba): move *LowLevelClient code into resource_*.go files
