+++ pkg/aruba/resource_cloud_server.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/Arubacloud/sdk-go/internal/clients/compute"
+	"github.com/Arubacloud/sdk-go/internal/restclient"
 	"github.com/Arubacloud/sdk-go/pkg/types"
 )
 
@@ -353,4 +355,213 @@ func cloudServerDerefString(p *string) string {
 		return ""
 	}
 	return *p
+}
+
+// ---------------------------------------------------------------------------
+// Low-level client seam + adapter
+// ---------------------------------------------------------------------------
+
+// cloudServerActions is an internal interface satisfied by cloudServersClientAdapter. It
+// allows *CloudServer to dispatch PowerOn/PowerOff/SetPassword without leaking the adapter
+// into the public API.
+type cloudServerActions interface {
+	powerOn(ctx context.Context, projectID, cloudServerID string, rp *types.RequestParameters) (*types.Response[types.CloudServerResponse], error)
+	powerOff(ctx context.Context, projectID, cloudServerID string, rp *types.RequestParameters) (*types.Response[types.CloudServerResponse], error)
+	setPassword(ctx context.Context, projectID, cloudServerID, password string, rp *types.RequestParameters) (*types.Response[any], error)
+}
+
+type cloudServerLowLevelClient interface {
+	List(ctx context.Context, projectID string, params *types.RequestParameters) (*types.Response[types.CloudServerList], error)
+	Get(ctx context.Context, projectID, cloudServerID string, params *types.RequestParameters) (*types.Response[types.CloudServerResponse], error)
+	Create(ctx context.Context, projectID string, body types.CloudServerRequest, params *types.RequestParameters) (*types.Response[types.CloudServerResponse], error)
+	Update(ctx context.Context, projectID, cloudServerID string, body types.CloudServerRequest, params *types.RequestParameters) (*types.Response[types.CloudServerResponse], error)
+	Delete(ctx context.Context, projectID, cloudServerID string, params *types.RequestParameters) (*types.Response[any], error)
+	PowerOn(ctx context.Context, projectID, cloudServerID string, params *types.RequestParameters) (*types.Response[types.CloudServerResponse], error)
+	PowerOff(ctx context.Context, projectID, cloudServerID string, params *types.RequestParameters) (*types.Response[types.CloudServerResponse], error)
+	SetPassword(ctx context.Context, projectID, cloudServerID string, body types.CloudServerPasswordRequest, params *types.RequestParameters) (*types.Response[any], error)
+}
+
+type cloudServersClientAdapter struct{ low cloudServerLowLevelClient }
+
+var _ cloudServerActions = (*cloudServersClientAdapter)(nil)
+
+func newCloudServersClientAdapter(rest *restclient.Client) *cloudServersClientAdapter {
+	if rest == nil {
+		return &cloudServersClientAdapter{}
+	}
+	return &cloudServersClientAdapter{low: compute.NewCloudServersClientImpl(rest)}
+}
+
+func (a *cloudServersClientAdapter) Create(ctx context.Context, cs *CloudServer, opts ...CallOption) (*CloudServer, error) {
+	if err := cs.Err(); err != nil {
+		return cs, err
+	}
+	if cs.ProjectID() == "" {
+		return cs, fmt.Errorf("Create: CloudServer has no parent project — call IntoProject first")
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Create(ctx, cs.ProjectID(), cs.toRequest(), rp)
+	populateHTTPEnvelope(&cs.httpEnvelopeMixin, resp)
+	if resp != nil && resp.Data != nil {
+		cs.fromResponse(resp.Data)
+	}
+	cs.actions = a
+	if err != nil {
+		return cs, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return cs, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return cs, nil
+}
+
+func (a *cloudServersClientAdapter) Update(ctx context.Context, cs *CloudServer, opts ...CallOption) (*CloudServer, error) {
+	if err := cs.Err(); err != nil {
+		return cs, err
+	}
+	if cs.CloudServerID() == "" {
+		return cs, fmt.Errorf("Update: CloudServer has no ID")
+	}
+	if cs.ProjectID() == "" {
+		return cs, fmt.Errorf("Update: CloudServer has no parent project — call IntoProject first")
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Update(ctx, cs.ProjectID(), cs.CloudServerID(), cs.toRequest(), rp)
+	populateHTTPEnvelope(&cs.httpEnvelopeMixin, resp)
+	if resp != nil && resp.Data != nil {
+		cs.fromResponse(resp.Data)
+	}
+	cs.actions = a
+	if err != nil {
+		return cs, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return cs, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return cs, nil
+}
+
+func (a *cloudServersClientAdapter) Get(ctx context.Context, ref Ref, opts ...CallOption) (*CloudServer, error) {
+	projectID, cloudServerID, err := cloudServerIDsFromRef(ref)
+	if err != nil {
+		return nil, err
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Get(ctx, projectID, cloudServerID, rp)
+	out := &CloudServer{}
+	populateHTTPEnvelope(&out.httpEnvelopeMixin, resp)
+	if resp != nil && resp.Data != nil {
+		out.fromResponse(resp.Data)
+	}
+	if out.projectID == "" {
+		out.projectID = projectID
+	}
+	out.actions = a
+	if err != nil {
+		return out, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return out, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return out, nil
+}
+
+func (a *cloudServersClientAdapter) Delete(ctx context.Context, ref Ref, opts ...CallOption) error {
+	projectID, cloudServerID, err := cloudServerIDsFromRef(ref)
+	if err != nil {
+		return err
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Delete(ctx, projectID, cloudServerID, rp)
+	if err != nil {
+		return err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return nil
+}
+
+func (a *cloudServersClientAdapter) List(ctx context.Context, project Ref, opts ...CallOption) (*List[*CloudServer], error) {
+	projectID, err := projectIDFromRef(project)
+	if err != nil {
+		return nil, err
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.List(ctx, projectID, rp)
+	if err != nil {
+		return nil, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return nil, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	var items []*CloudServer
+	if resp != nil && resp.Data != nil {
+		items = make([]*CloudServer, 0, len(resp.Data.Values))
+		for i := range resp.Data.Values {
+			cs := &CloudServer{}
+			cs.fromResponse(&resp.Data.Values[i])
+			if cs.projectID == "" {
+				cs.projectID = projectID
+			}
+			cs.actions = a
+			items = append(items, cs)
+		}
+	}
+	refetch := func(_ context.Context, _ string) (*List[*CloudServer], error) {
+		return nil, fmt.Errorf("List pagination by URL not yet wired; re-call List with adjusted CallOptions")
+	}
+	var total int64
+	var self, prev, next, first, last string
+	if resp != nil && resp.Data != nil {
+		total = resp.Data.Total
+		self = resp.Data.Self
+		prev = resp.Data.Prev
+		next = resp.Data.Next
+		first = resp.Data.First
+		last = resp.Data.Last
+	}
+	return newList(items, total, self, prev, next, first, last, resp, opts, refetch), nil
+}
+
+// Internal action methods — satisfy cloudServerActions; called by *CloudServer action methods.
+
+func (a *cloudServersClientAdapter) powerOn(ctx context.Context, projectID, cloudServerID string, rp *types.RequestParameters) (*types.Response[types.CloudServerResponse], error) {
+	return a.low.PowerOn(ctx, projectID, cloudServerID, rp)
+}
+
+func (a *cloudServersClientAdapter) powerOff(ctx context.Context, projectID, cloudServerID string, rp *types.RequestParameters) (*types.Response[types.CloudServerResponse], error) {
+	return a.low.PowerOff(ctx, projectID, cloudServerID, rp)
+}
+
+func (a *cloudServersClientAdapter) setPassword(ctx context.Context, projectID, cloudServerID, password string, rp *types.RequestParameters) (*types.Response[any], error) {
+	return a.low.SetPassword(ctx, projectID, cloudServerID, types.CloudServerPasswordRequest{Password: password}, rp)
+}
+
+// cloudServerIDsFromRef extracts (projectID, cloudServerID) from a Ref.
+func cloudServerIDsFromRef(ref Ref) (projectID, cloudServerID string, err error) {
+	csID, ok := extractID(ref, func(r Ref) (string, bool) {
+		if w, ok := r.(withCloudServerID); ok {
+			return w.CloudServerID(), true
+		}
+		return "", false
+	}, "cloudServers")
+	if !ok || csID == "" {
+		return "", "", fmt.Errorf("cannot determine CloudServer ID from Ref %q", ref.URI())
+	}
+	pid, ok := extractID(ref, func(r Ref) (string, bool) {
+		if w, ok := r.(withProjectID); ok {
+			return w.ProjectID(), true
+		}
+		return "", false
+	}, "projects")
+	if !ok || pid == "" {
+		return "", "", fmt.Errorf("cannot determine project ID from Ref %q", ref.URI())
+	}
+	return pid, csID, nil
 }
