+++ pkg/aruba/resource_snapshot.go
@@ -1,8 +1,11 @@
 package aruba
 
 import (
+	"context"
 	"fmt"
 
+	"github.com/Arubacloud/sdk-go/internal/clients/storage"
+	"github.com/Arubacloud/sdk-go/internal/restclient"
 	"github.com/Arubacloud/sdk-go/pkg/types"
 )
 
@@ -160,4 +163,183 @@ func snapshotDerefString(p *string) string {
 		return ""
 	}
 	return *p
+}
+
+// ---------------------------------------------------------------------------
+// Low-level client interface, adapter, constructor, and methods
+// ---------------------------------------------------------------------------
+
+type snapshotLowLevelClient interface {
+	List(ctx context.Context, projectID string, params *types.RequestParameters) (*types.Response[types.SnapshotList], error)
+	Get(ctx context.Context, projectID, snapshotID string, params *types.RequestParameters) (*types.Response[types.SnapshotResponse], error)
+	Create(ctx context.Context, projectID string, body types.SnapshotRequest, params *types.RequestParameters) (*types.Response[types.SnapshotResponse], error)
+	Update(ctx context.Context, projectID, snapshotID string, body types.SnapshotRequest, params *types.RequestParameters) (*types.Response[types.SnapshotResponse], error)
+	Delete(ctx context.Context, projectID, snapshotID string, params *types.RequestParameters) (*types.Response[any], error)
+}
+
+type snapshotsClientAdapter struct{ low snapshotLowLevelClient }
+
+func newSnapshotsClientAdapter(rest *restclient.Client) *snapshotsClientAdapter {
+	if rest == nil {
+		return &snapshotsClientAdapter{}
+	}
+	return &snapshotsClientAdapter{
+		low: storage.NewSnapshotsClientImpl(rest, storage.NewVolumesClientImpl(rest)),
+	}
+}
+
+func (a *snapshotsClientAdapter) Create(ctx context.Context, snap *Snapshot, opts ...CallOption) (*Snapshot, error) {
+	if err := snap.Err(); err != nil {
+		return snap, err
+	}
+	if snap.ProjectID() == "" {
+		return snap, fmt.Errorf("Create: Snapshot has no project — call IntoProject first")
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Create(ctx, snap.ProjectID(), snap.toRequest(), rp)
+	populateHTTPEnvelope(&snap.httpEnvelopeMixin, resp)
+	if resp != nil && resp.Data != nil {
+		snap.fromResponse(resp.Data)
+	}
+	if err != nil {
+		return snap, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return snap, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return snap, nil
+}
+
+func (a *snapshotsClientAdapter) Get(ctx context.Context, ref Ref, opts ...CallOption) (*Snapshot, error) {
+	projectID, snapshotID, err := snapshotIDsFromRef(ref)
+	if err != nil {
+		return nil, err
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Get(ctx, projectID, snapshotID, rp)
+	out := &Snapshot{}
+	populateHTTPEnvelope(&out.httpEnvelopeMixin, resp)
+	if resp != nil && resp.Data != nil {
+		out.fromResponse(resp.Data)
+	}
+	if out.projectID == "" {
+		out.projectID = projectID
+	}
+	if err != nil {
+		return out, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return out, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return out, nil
+}
+
+func (a *snapshotsClientAdapter) Update(ctx context.Context, snap *Snapshot, opts ...CallOption) (*Snapshot, error) {
+	if err := snap.Err(); err != nil {
+		return snap, err
+	}
+	if snap.ID() == "" {
+		return snap, fmt.Errorf("Update: Snapshot has no ID — call Get first or seed from response metadata")
+	}
+	if snap.ProjectID() == "" {
+		return snap, fmt.Errorf("Update: Snapshot has no project — call IntoProject first")
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Update(ctx, snap.ProjectID(), snap.ID(), snap.toRequest(), rp)
+	populateHTTPEnvelope(&snap.httpEnvelopeMixin, resp)
+	if resp != nil && resp.Data != nil {
+		snap.fromResponse(resp.Data)
+	}
+	if err != nil {
+		return snap, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return snap, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return snap, nil
+}
+
+func (a *snapshotsClientAdapter) Delete(ctx context.Context, ref Ref, opts ...CallOption) error {
+	projectID, snapshotID, err := snapshotIDsFromRef(ref)
+	if err != nil {
+		return err
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Delete(ctx, projectID, snapshotID, rp)
+	if err != nil {
+		return err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return nil
+}
+
+func (a *snapshotsClientAdapter) List(ctx context.Context, project Ref, opts ...CallOption) (*List[*Snapshot], error) {
+	projectID, err := projectIDFromRef(project)
+	if err != nil {
+		return nil, err
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.List(ctx, projectID, rp)
+	if err != nil {
+		return nil, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return nil, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	var items []*Snapshot
+	if resp != nil && resp.Data != nil {
+		items = make([]*Snapshot, 0, len(resp.Data.Values))
+		for i := range resp.Data.Values {
+			snap := &Snapshot{}
+			snap.fromResponse(&resp.Data.Values[i])
+			if snap.projectID == "" {
+				snap.projectID = projectID
+			}
+			items = append(items, snap)
+		}
+	}
+	refetch := func(_ context.Context, _ string) (*List[*Snapshot], error) {
+		return nil, fmt.Errorf("List pagination by URL not yet wired; re-call List with adjusted CallOptions")
+	}
+	var total int64
+	var self, prev, next, first, last string
+	if resp != nil && resp.Data != nil {
+		total = resp.Data.Total
+		self = resp.Data.Self
+		prev = resp.Data.Prev
+		next = resp.Data.Next
+		first = resp.Data.First
+		last = resp.Data.Last
+	}
+	return newList(items, total, self, prev, next, first, last, resp, opts, refetch), nil
+}
+
+// snapshotIDsFromRef extracts (projectID, snapshotID) from a Ref.
+func snapshotIDsFromRef(ref Ref) (projectID, snapshotID string, err error) {
+	sid, ok := extractID(ref, func(r Ref) (string, bool) {
+		if w, ok := r.(withSnapshotID); ok {
+			return w.SnapshotID(), true
+		}
+		return "", false
+	}, "snapshots")
+	if !ok || sid == "" {
+		return "", "", fmt.Errorf("cannot determine Snapshot ID from Ref %q", ref.URI())
+	}
+	pid, ok := extractID(ref, func(r Ref) (string, bool) {
+		if w, ok := r.(withProjectID); ok {
+			return w.ProjectID(), true
+		}
+		return "", false
+	}, "projects")
+	if !ok || pid == "" {
+		return "", "", fmt.Errorf("cannot determine project ID from Ref %q", ref.URI())
+	}
+	return pid, sid, nil
 }
