+++ pkg/aruba/resource_storage_restore.go
@@ -1,8 +1,11 @@
 package aruba
 
 import (
+	"context"
 	"fmt"
 
+	"github.com/Arubacloud/sdk-go/internal/clients/storage"
+	"github.com/Arubacloud/sdk-go/internal/restclient"
 	"github.com/Arubacloud/sdk-go/pkg/types"
 )
 
@@ -115,4 +118,203 @@ func storageRestoreDerefString(p *string) string {
 		return ""
 	}
 	return *p
+}
+
+// ---------------------------------------------------------------------------
+// Low-level client interface, adapter, constructor, and methods
+// ---------------------------------------------------------------------------
+
+type storageRestoreLowLevelClient interface {
+	List(ctx context.Context, projectID, backupID string, params *types.RequestParameters) (*types.Response[types.StorageRestoreList], error)
+	Get(ctx context.Context, projectID, backupID, restoreID string, params *types.RequestParameters) (*types.Response[types.StorageRestoreResponse], error)
+	Create(ctx context.Context, projectID, backupID string, body types.StorageRestoreRequest, params *types.RequestParameters) (*types.Response[types.StorageRestoreResponse], error)
+	Update(ctx context.Context, projectID, backupID, restoreID string, body types.StorageRestoreRequest, params *types.RequestParameters) (*types.Response[types.StorageRestoreResponse], error)
+	Delete(ctx context.Context, projectID, backupID, restoreID string, params *types.RequestParameters) (*types.Response[any], error)
+}
+
+type storageRestoresClientAdapter struct{ low storageRestoreLowLevelClient }
+
+func newStorageRestoresClientAdapter(rest *restclient.Client) *storageRestoresClientAdapter {
+	if rest == nil {
+		return &storageRestoresClientAdapter{}
+	}
+	// NewRestoreClientImpl panics if backupClient is nil — instantiate one internally
+	// so the public adapter constructor stays single-arg.
+	return &storageRestoresClientAdapter{
+		low: storage.NewRestoreClientImpl(rest, storage.NewBackupClientImpl(rest)),
+	}
+}
+
+func (a *storageRestoresClientAdapter) Create(ctx context.Context, r *StorageRestore, opts ...CallOption) (*StorageRestore, error) {
+	if err := r.Err(); err != nil {
+		return r, err
+	}
+	if r.BackupID() == "" || r.ProjectID() == "" {
+		return r, fmt.Errorf("Create: StorageRestore has no parent backup — call IntoBackup first")
+	}
+	if r.targetRef == nil {
+		return r, fmt.Errorf("Create: StorageRestore has no target — call WithTarget first")
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Create(ctx, r.ProjectID(), r.BackupID(), r.toRequest(), rp)
+	populateHTTPEnvelope(&r.httpEnvelopeMixin, resp)
+	if resp != nil && resp.Data != nil {
+		r.fromResponse(resp.Data)
+	}
+	if err != nil {
+		return r, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return r, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return r, nil
+}
+
+func (a *storageRestoresClientAdapter) Get(ctx context.Context, ref Ref, opts ...CallOption) (*StorageRestore, error) {
+	projectID, backupID, restoreID, err := restoreIDsFromRef(ref)
+	if err != nil {
+		return nil, err
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Get(ctx, projectID, backupID, restoreID, rp)
+	out := &StorageRestore{}
+	populateHTTPEnvelope(&out.httpEnvelopeMixin, resp)
+	if resp != nil && resp.Data != nil {
+		out.fromResponse(resp.Data)
+	}
+	if out.projectID == "" {
+		out.projectID = projectID
+	}
+	if out.backupID == "" {
+		out.backupID = backupID
+	}
+	if err != nil {
+		return out, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return out, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return out, nil
+}
+
+func (a *storageRestoresClientAdapter) Update(ctx context.Context, r *StorageRestore, opts ...CallOption) (*StorageRestore, error) {
+	if err := r.Err(); err != nil {
+		return r, err
+	}
+	if r.ID() == "" {
+		return r, fmt.Errorf("Update: StorageRestore has no ID — call Get first or seed from response metadata")
+	}
+	if r.BackupID() == "" || r.ProjectID() == "" {
+		return r, fmt.Errorf("Update: StorageRestore has no parent backup — call IntoBackup first")
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Update(ctx, r.ProjectID(), r.BackupID(), r.ID(), r.toRequest(), rp)
+	populateHTTPEnvelope(&r.httpEnvelopeMixin, resp)
+	if resp != nil && resp.Data != nil {
+		r.fromResponse(resp.Data)
+	}
+	if err != nil {
+		return r, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return r, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return r, nil
+}
+
+func (a *storageRestoresClientAdapter) Delete(ctx context.Context, ref Ref, opts ...CallOption) error {
+	projectID, backupID, restoreID, err := restoreIDsFromRef(ref)
+	if err != nil {
+		return err
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.Delete(ctx, projectID, backupID, restoreID, rp)
+	if err != nil {
+		return err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	return nil
+}
+
+func (a *storageRestoresClientAdapter) List(ctx context.Context, backup Ref, opts ...CallOption) (*List[*StorageRestore], error) {
+	projectID, backupID, err := backupIDsFromRef(backup)
+	if err != nil {
+		return nil, err
+	}
+	co := applyCallOptions(opts)
+	rp := co.toRequestParameters()
+	resp, err := a.low.List(ctx, projectID, backupID, rp)
+	if err != nil {
+		return nil, err
+	}
+	if resp != nil && !resp.IsSuccess() {
+		return nil, &HTTPError{StatusCode: resp.StatusCode, Body: resp.RawBody, ErrResp: resp.Error}
+	}
+	var items []*StorageRestore
+	if resp != nil && resp.Data != nil {
+		items = make([]*StorageRestore, 0, len(resp.Data.Values))
+		for i := range resp.Data.Values {
+			v := &StorageRestore{}
+			v.fromResponse(&resp.Data.Values[i])
+			if v.projectID == "" {
+				v.projectID = projectID
+			}
+			if v.backupID == "" {
+				v.backupID = backupID
+			}
+			items = append(items, v)
+		}
+	}
+	refetch := func(_ context.Context, _ string) (*List[*StorageRestore], error) {
+		return nil, fmt.Errorf("List pagination by URL not yet wired; re-call List with adjusted CallOptions")
+	}
+	var total int64
+	var self, prev, next, first, last string
+	if resp != nil && resp.Data != nil {
+		total = resp.Data.Total
+		self = resp.Data.Self
+		prev = resp.Data.Prev
+		next = resp.Data.Next
+		first = resp.Data.First
+		last = resp.Data.Last
+	}
+	return newList(items, total, self, prev, next, first, last, resp, opts, refetch), nil
+}
+
+// restoreIDsFromRef extracts (projectID, backupID, restoreID) from a Ref.
+func restoreIDsFromRef(ref Ref) (projectID, backupID, restoreID string, err error) {
+	rid, ok := extractID(ref, func(r Ref) (string, bool) {
+		if w, ok := r.(withRestoreID); ok {
+			return w.RestoreID(), true
+		}
+		return "", false
+	}, "restores")
+	if !ok || rid == "" {
+		return "", "", "", fmt.Errorf("cannot determine StorageRestore ID from Ref %q", ref.URI())
+	}
+	bid, ok := extractID(ref, func(r Ref) (string, bool) {
+		if w, ok := r.(withBackupID); ok {
+			return w.BackupID(), true
+		}
+		return "", false
+	}, "backups")
+	if !ok || bid == "" {
+		return "", "", "", fmt.Errorf("cannot determine backup ID from Ref %q", ref.URI())
+	}
+	pid, ok := extractID(ref, func(r Ref) (string, bool) {
+		if w, ok := r.(withProjectID); ok {
+			return w.ProjectID(), true
+		}
+		return "", false
+	}, "projects")
+	if !ok || pid == "" {
+		return "", "", "", fmt.Errorf("cannot determine project ID from Ref %q", ref.URI())
+	}
+	return pid, bid, rid, nil
 }
